reverted the script chanegs
