+++ blocks/product-hero/product-hero.js
@@ -5,7 +5,7 @@ import {
   getAuthorization, getCommerceBase,
   getProductResponse,
 } from '../../scripts/commerce.js';
-import { decorateModals, createOptimizedS7Picture } from '../../scripts/scripts.js';
+import { createOptimizedS7Picture, decorateModals } from '../../scripts/scripts.js';
 
 function showImage(e) {
   const selectedImage = document.querySelector('.image-content picture');
